+++ packages/jsonapi/src/RequestHandling/ExpectedPropertyCollection.php
@@ -30,21 +30,25 @@ class ExpectedPropertyCollection
     ) {}
 
     /**
+     * @param int<0, 8192> $validationLevelDepth
+     *
      * @return array<non-empty-string, list<Constraint>>
      */
-    public function getRequiredAttributes(): array
+    public function getRequiredAttributes(int $validationLevelDepth, bool $allowAnythingBelowDepth): array
     {
-        return array_fill_keys($this->requiredAttributes, $this->getConstraintsForAttribute());
+        return array_fill_keys($this->requiredAttributes, $this->getConstraintsForAttribute($validationLevelDepth, $allowAnythingBelowDepth));
     }
 
     /**
+     * @param int<0, 8192> $validationLevelDepth
+     *
      * @return array<non-empty-string, list<Constraint>>
      */
-    public function getAllowedAttributes(): array
+    public function getAllowedAttributes(int $validationLevelDepth, bool $allowAnythingBelowDepth): array
     {
         return array_fill_keys(
             array_merge($this->requiredAttributes, $this->optionalAttributes),
-            $this->getConstraintsForAttribute()
+            $this->getConstraintsForAttribute($validationLevelDepth, $allowAnythingBelowDepth)
         );
     }
 
@@ -81,16 +85,56 @@ class ExpectedPropertyCollection
     }
 
     /**
+     * Ensures attributes are either a primitive type, null or (optionally) an array.
+     *
+     * @param int<0, 8192> $validationLevelDepth The number of levels for which the returned constraint will ensure valid attribute values.
+     *                                           If `1`, the returned constraint will ensure that the attribute value itself is
+     *                                           a primitive type, `null` or an array. The higher the depth value, the more validation is
+     *                                           done for nested arrays.
+     * @param bool $allowAnythingBelowDepth Determines if arrays are allowed at all below the validation depth. E.g. if
+     *                                      this parameter is set to `false` and the depth parameter is set to `1`,
+     *                                      then an attribute value is allowed to be an array, but such array can not
+     *                                      contain other arrays, but primitive types or `null` instead only. If set to
+     *                                      `true` and `1`, then an attribute value that is an array may contain anything.
+     *
      * @return list<Constraint>
      */
-    protected function getConstraintsForAttribute(): array
+    protected function getConstraintsForAttribute(int $validationLevelDepth, bool $allowAnythingBelowDepth): array
+    {
+        $lowestLevelAllowedTypes = ['string', 'int', 'float', 'bool'];
+        if ($allowAnythingBelowDepth) {
+            $lowestLevelAllowedTypes[] = 'array';
+        }
+
+        $resultConstraints = [new Assert\Type($lowestLevelAllowedTypes)];
+        for ($i = 0; $i < $validationLevelDepth; $i++) {
+            $resultConstraints = $this->createAttributeConstraintLevel($resultConstraints);
+        }
+
+        return $resultConstraints;
+    }
+
+    /**
+     * @param list<Constraint> $lowerLevel
+     *
+     * @return list<Constraint>
+     */
+    protected function createAttributeConstraintLevel(array $lowerLevel): array
     {
         return [
             new Assert\AtLeastOneOf([
-                // primitive values are always valid
-                new Assert\Type(['string', 'int', 'float', 'bool'], 'If not null, attributes must be of type {{ type }}.'),
+                // primitive values are always valid (null is implicitly allowed)
+                new Assert\Type(['string', 'int', 'float', 'bool'], 'If not null or array, attributes must be of type {{ type }}.'),
                 // if the type is an array, then it must contain only non-null strings
-                new class extends Assert\Compound {
+                new class($lowerLevel) extends Assert\Compound {
+                    /**
+                     * @param list<Constraint> $lowerLevel
+                     */
+                    public function __construct(protected readonly array $lowerLevel)
+                    {
+                        parent::__construct();
+                    }
+
                     /**
                      * @inheritDoc
                      */
@@ -98,15 +142,11 @@ class ExpectedPropertyCollection
                     {
                         return [
                             new Assert\Type('array'),
-                            new Assert\All([
-                                new Assert\Type('string'),
-                                new Assert\NotNull(),
-                            ]),
+                            new Assert\All($this->lowerLevel),
                         ];
                     }
                 }
-                // TODO: the JSON:API specification may allow more attribute types, especially regarding `array`, but those require more complex validation
-            ]),
+            ])
         ];
     }